wip: ritual dagger stubs

+++ src/AllBeginningsMod/Utilities/_Utilities/Helper.cs
@@ -56,7 +56,7 @@ internal static partial class Helper {
     }
 
     public static Point ToSafeTileCoordinates(this Vector2 vec) {
-        return new Point(MathHelper.Clamp((int)vec.X >> 4, 0, Main.maxTilesX), MathHelper.Clamp((int)vec.Y >> 4, 0, Main.maxTilesY));
+        return new Point((int)MathHelper.Clamp((int)vec.X >> 4, 0, Main.maxTilesX), (int)MathHelper.Clamp((int)vec.Y >> 4, 0, Main.maxTilesY));
     }
 
     public static Vector2 InverseKinematic(Vector2 start, Vector2 end, float A, float B, bool flip) {
